refactor(agentic): split branch-lifecycle into derive/store/actions
Single ~620-line module fanned out into three slices behind a
re-export barrel so the public import surface (./branch-lifecycle)
stays unchanged:

  branch-lifecycle-derive   pure deriveLifecycle + types; takes the
                            resolved paneAgentState as an input
                            instead of reading paneAgentStateStore
                            and the test-override map at derivation
                            time, removing the only impurity in the
                            old derive function.
  branch-lifecycle-store    registry, sync, recompute loop, mutators,
                            test helpers; owns the override map and
                            resolves paneId → AgentState before
                            invoking derive.
  branch-lifecycle-actions  markAbandoned only.

markAbandoned previously cast `{ id }` as Workspace when the
workspace had already been removed, then called confirmAndCloseWorkspace
with a fake object so the unit test could observe the call. Replaced
with an honest early-return: no workspace, no close. The original
test now wires the workspace into the mocked workspaces store so the
production path runs end-to-end, and a new test pins the early-return
behavior for the workspace-gone case.
